requirements.txt
tornado==4.4.1
# jinja2==2.7.3
lxml==3.4.4
# markupsafe==0.23
# matplotlib==1.4.3
# pip install -i http://pypi.douban.com/simple/
numpy==1.9.2
pillow==2.9.0
# pyparsing==2.0.3 # 代码整理工具
# python-dateutil==2.4.2
# scipy==0.15.1 # 数学计算工具
# setuptools==18.0.1
# six==1.9.0 # 代码比对工具
tinydb==2.3.2
# toro==0.8 #  tornado 的队列工具
pyyaml==3.12
redis==2.10.5

